Deduplicate pagination and tooltip logic from PlaceholderPanel/PlaceholderListPanel into AbstractPanel
Move identical createNextButton, createPreviousButton, and addTooltips implementations
up into AbstractPanel with hook methods getPagedItemCount() and getPagedItemType() so
subclasses only supply their list size and item-type key.
